+++ playwright/tests/office/qaecsr/qaeFlows.spec.js
@@ -5,7 +5,7 @@
  */
 
 // @ts-check
-import { test, expect, OfficePage } from '../../utils/officeTest';
+import { test, expect, OfficePage } from '../../utils/office/officeTest';
 
 /**
  * QaeFlowPage test fixture
