+++ playwright/tests/office/txo/sitUpdates.spec.js
@@ -1,5 +1,5 @@
 // @ts-check
-import { test, expect } from '../../utils/officeTest';
+import { test, expect } from '../../utils/office/officeTest';
 
 import { TooFlowPage } from './tooTestFixture';
